requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
networkx>=2.6.0
tqdm>=4.62.0
torch>=1.9.0

